refactor: main파일의 MenuHandler를 BankingProgramHandler으로 변경하고 파일을 분리 #4
- GuideDescription 열거형 추가
- Menu 열거형 description 프로퍼티 추가
- startToReceiveInput 메서드 default 수정
