--- BankManager.swift
@@ -1,2 +1,0 @@
-
-import Foundation
