+++ BankManagerConsoleApp/BankManagerConsoleApp/Bank.swift
