+++ BankManagerConsoleApp/BankManagerConsoleApp/Error/LinkedListError.swift
