+++ BankManagerConsoleApp/BankManagerConsoleApp/LinkedList.swift
